+++ src/hooks/login/index.ts
@@ -2,5 +2,6 @@ export { useExtensionLogin } from './useExtensionLogin';
 export { useGetModalLoginMethods } from './useGetModalLoginMethods';
 export { useLedgerLogin } from './useLedgerLogin';
 export { useOperaLogin } from './useOperaLogin';
+export { useXaliasLogin } from './useXaliasLogin';
 export { useWalletConnectV2Login } from './useWalletConnectV2Login';
 export { useWebWalletLogin } from './useWebWalletLogin';